Add sample FPS adapter to support flexible way to reduce FPS.
1. Add sample FPS adapter to support flexible way to reduce FPS by using
a bit pattern and period to decide which frame should be kept.
2. Remove workaround code for TinyViz.

+++ tests/sample/source/SampleCamera.cpp
@@ -155,7 +155,7 @@ RideHalError_e SampleCamera::Init( std::string name, SampleConfig_t &config )
             }
         }
 
-        m_camConfig.camFrameDropPat = Get( config, "frame_drop_patten", 0 );
+        m_camConfig.camFrameDropPat = Get( config, "frame_drop_patten", (uint32_t)0 );
 
         m_camConfig.opMode = Get( config, "op_mode", (uint32_t) QCARCAM_OPMODE_OFFLINE_ISP );
